+++ src/data/emacs-mode/agda2-abbrevs.el
@@ -1,4 +1,5 @@
 ;; agda2-abbrevs.el --- Default Agda abbrevs
+;; SPDX-License-Identifier: MIT License
 
 ;;; Commentary: