+++ CHANGELOG.md
@@ -154,6 +154,76 @@ Language
 * Binary integer literals with prefix `0b` (for instance, `0b11001001`) are now
   supported.
 
+Builtins
+--------
+
+- Primitive operations for floating-point numbers changed. The equalities now
+  follow IEEE 754 equality, after unifying all NaNs. Primitive inequality was
+  added:
+  ```agda
+  primFloatEquality   : Float -> Float -> Bool -- from primFloatNumericEquality
+  primFloatLess       : Float -> Float -> Bool -- from primFloatNumericLess
+  primFloatInequality : Float -> Float -> Bool -- new
+  ```
+  The “numeric” relations are now deprecated.
+
+  There are several new predicates on floating-point numbers:
+  ```agda
+  primFloatIsInfinite     : Float -> Bool -- new
+  primFloatIsNaN          : Float -> Bool -- new
+  primFloatIsSafeInteger  : Float -> Bool -- new
+  ```
+  The primFloatIsSafeInteger function determines whether the value is a number
+  that is a safe integer, i.e., is within the range where the arithmetic
+  operations do not lose precision.
+
+  The operations for conversion to integers (primRound, primFloor, and
+  primCeiling) were renamed for consistency, and return a value of type `Maybe
+  Int`, returning `nothing` for NaN and the infinities:
+  ```agda
+  primFloatRound   : Float → Maybe Int -- from primRound
+  primFloatFloor   : Float → Maybe Int -- from primFloor
+  primFloatCeiling : Float → Maybe Int -- from primCeiling
+  ```
+
+  There are several new conversions:
+  ```agda
+  primIntToFloat    : Int -> Float               -- new
+  primFloatToRatio  : Float -> (Int × Nat)       -- new
+  primRatioToFloat  : Int -> Nat -> Float        -- new
+  primFloatDecode   : Float -> Maybe (Int × Int) -- new
+  primFloatEncode   : Int -> Int -> Maybe Float  -- new
+  ```
+  The `primFloatDecode` function decodes a floating-point number f to a mantissa
+  and exponent, such that `f = mantissa * 2 ^ exponent`, normalised such that
+  the mantissa is the smallest possible number. The `primFloatEncode` function
+  encodes a pair of a mantissa and exponent to a floating-point number.
+
+  There are several new operations:
+  ```agda
+  primFloatPow        : Float -> Float -> Float -- new
+  primFloatATan2      : Float -> Float -> Float -- from primATan2
+  primFloatSinh       : Float -> Float          -- new
+  primFloatCosh       : Float -> Float          -- new
+  primFloatTanh       : Float -> Float          -- new
+  primFloatASinh      : Float -> Float          -- new
+  primFloatACosh      : Float -> Float          -- new
+  primFloatATanh      : Float -> Float          -- new
+  ```
+  Furthermore, the following operations were renamed for consistency:
+  ```agda
+  primFloatExp        : Float -> Float          -- from primExp
+  primFloatSin        : Float -> Float          -- from primSin
+  primFloatLog        : Float -> Float          -- from primLog
+  primFloatCos        : Float -> Float          -- from primCos
+  primFloatTan        : Float -> Float          -- from primTan
+  primFloatASin       : Float -> Float          -- from primASin
+  primFloatACos       : Float -> Float          -- from primACos
+  primFloatATan       : Float -> Float          -- from primATan
+  ```
+
+  All of these operations are implemented on the JavaScript backend.
+
 Reflection
 ----------
 
@@ -238,6 +308,8 @@ Emacs mode
   ```
 
   `C-u C-u C-u C-c C-n downFrom 5` returns `4 ∷ downFrom 4`.
+
+* New keyboard shortcut `C-c C-x C-i` for toggling display of irrelevant arguments.
 
 
 JS backend